Add files via upload
Step 1: retrieve input from the user that should be an array of type int
Step 2: next, derive the first two elements of the array and return the difference(if the difference is negative, return a positive integer). Let the first element be a and the second be b and the difference returned as v1
Step 3 : after step 2 remove the first two elements, do check if the array is empty at this point and find the smallest array element save it as s and larger array element as l(Max_value used here) and now return the difference between these two element as v2(abs).
Step 4: return the sum of v1 and v2 as max profit(also static variable in this case).